passage au séjour
git-svn-id: svn://svn.code.sf.net/p/mediboard/code/trunk@111 a21d6be2-e512-0410-a060-91f5a2946631

+++ modules/dPplanningOp/do_planning_aed.php
@@ -37,7 +37,8 @@ if (intval(dPgetParam($_POST, 'del'))) {
   }
   $sejour->bindToOp($do->_obj->operation_id);
   $do->doDelete();
-  $sejour->delete();
+  if($sejour->sejour_id)
+    $sejour->delete();
 } else {
   if($do->_obj->plageop_id && $do->_obj->pat_id) {
     $do->modifyMsg = "Opération modifiée";
@@ -51,13 +52,15 @@ if (intval(dPgetParam($_POST, 'del'))) {
   }
   $do->doStore();
   $sejour->bindToOp($do->_obj->operation_id);
-  $sejour->store();
   // Pour que la redirection prenne vraiment l'objet en compte et pas que les valeurs du POST
   // -> on reload l'objet
   $do->_obj->load($do->_obj->operation_id);
   // Stockage du séjour créé
-  $do->_obj->sejour_id = $sejour->sejour_id;
-  $do->_obj->store();
+  if($do->_obj->pat_id) {
+    $sejour->store();
+    $do->_obj->sejour_id = $sejour->sejour_id;
+    $do->_obj->store();
+  }
   if($otherm = dPgetParam( $_POST, 'otherm', 0))
     $m = $otherm;
   if($m == "dPhospi")
